refactor: ui, model, utils로 패키지 구분
