--- src/main/java/racinggame/NumberGenerator.java
@@ -1,5 +1,0 @@
-package racinggame;
-
-interface NumberGenerator {
-	int generate();
-}
