+++ src/main/java/racinggame/model/Car.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.model;
 
 public class Car implements Comparable<Car> {
 	private CarName name;
