+++ src/main/java/racinggame/model/CarMileage.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.model;
 
 import java.util.Objects;
