+++ src/main/java/racinggame/model/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.model;
 
 public class MoveStrategy {
 	private final int number;
