+++ src/main/java/racinggame/model/NumberGenerator.java
@@ -1,0 +1,5 @@
+package racinggame.model;
+
+public interface NumberGenerator {
+	int generate();
+}
