+++ src/main/java/racinggame/model/RacingGame.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.model;
 
 import java.util.List;
