+++ src/main/java/racinggame/model/RacingGameRule.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.model;
 
 public class RacingGameRule {
 	public static int MINIMUM_OF_GAME_NUMBER = 0;
