+++ src/test/java/racinggame/CarGeneratorTest.java
@@ -7,6 +7,9 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import racinggame.model.Car;
+import racinggame.utils.CarGenerator;
+
 public class CarGeneratorTest {
 	@Test
 	@DisplayName("이름으로 자동차 생성")
